+++ src/main/java/com/seancarroll/foundationdb/es/SetStreamMetadataResult.java
@@ -11,17 +11,17 @@ public class SetStreamMetadataResult {
     /**
      * The current version of the stream at the time the metadata was written.
      */
-    public final int currentVersion;
+    public final long currentVersion;
 
     /**
      * Initializes a new instance of the {@link SetStreamMetadataResult}.
      * @param currentVersion
      */
-    public SetStreamMetadataResult(int currentVersion) {
+    public SetStreamMetadataResult(long currentVersion) {
         this.currentVersion = currentVersion;
     }
 
-    public int getCurrentVersion() {
+    public long getCurrentVersion() {
         return currentVersion;
     }
